Move anti-forgery filters to the Nop.Web.Framework.Mvc.Filters folder
